% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GPUepiblaster.R
\name{GPUepiblaster1geno}
\alias{GPUepiblaster1geno}
\title{GPU-based calculation of the difference of correlation coefficients and compute Z test with one genotype input}
\usage{
GPUepiblaster1geno(geno, pheno, chunk = 1000, zpthres = 1e-05,
  outfile = "NONE", suffix = ".txt", gpuidx = 0, ...)
}
\arguments{
\item{geno}{is the normalized genotype data. It can be a matrix or a dataframe, or a big.matrix object from \pkg{bigmemory}.
The columns contain the information of variables and the rows contain the information of samples.}

\item{pheno}{a vector containing the normalized phenotype information. The values are either 0 or 1.}

\item{chunk}{is the number of variants in each chunk.}

\item{zpthres}{is the significance threshold to select variant pairs for output. The default is 1e-6.}

\item{outfile}{is the prefix of out filename.}

\item{suffix}{is the suffix of out filename.}

\item{gpuidx}{is the integer index of GPU(device) starting from 0. The default is 0.}

\item{...}{not used.}
}
\value{
null
}
\description{
Calculate the difference of correlation coefficients between cases and controls,
conduct Z test for the differences (values) and choose variant pairs with the significance below the given threshold for output.
}
\examples{
# simulate some data
set.seed(123)
geno1 <- matrix(sample(0:2, size = 1000, replace = TRUE, prob = c(0.5, 0.3, 0.2)), ncol = 10)
dimnames(geno1) <- list(row = paste0("IND", 1:nrow(geno1)), col = paste0("rs", 1:ncol(geno1)))
p1 <- c(rep(0, 60), rep(1, 40))

# normalized data
geno1 <- scale(geno1)

# one genotype with case-control phenotype
GPUepiblaster1geno(geno = geno1,
pheno = p1,
outfile = "episcan_1geno_cc",
suffix = ".txt",
zpthres = 0.9,
chunk = 10)

# take a look at the result
res <- read.table("episcan_1geno_cc.txt", header = TRUE, stringsAsFactors = FALSE)
head(res)
}
\author{
Beibei Jiang \email{beibei_jiang@psych.mpg.de}
}
