% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GPUepiHSIC.R
\name{GPUepiHSIC}
\alias{GPUepiHSIC}
\title{Calculate HSIC values on GPU}
\usage{
GPUepiHSIC(A = NULL, B = NULL, P = NULL, s = NULL, GPUidx = 0,
  ...)
}
\arguments{
\item{A}{is one matrix.}

\item{B}{is one matrix.}

\item{P}{is "phenoype", a vector.}

\item{s}{is the standard deviation of the epiHSIC distribution}

\item{GPUidx}{is the integer index of GPU(device) starting from 0. Default is 0.}

\item{...}{not used.}
}
\value{
a matrix
}
\description{
Calculate HSIC values on GPU
}
\author{
Beibei Jiang \email{beibei_jiang@psych.mpg.de}
}
