% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GPUepiblaster.R
\name{GPUgetcor}
\alias{GPUgetcor}
\title{Get correlation matrix}
\usage{
GPUgetcor(A = NULL, B = NULL, method = "pearson", GPUidx = 0, ...)
}
\arguments{
\item{A}{is a matrix or data.frame.}

\item{B}{is a matrix or data.frame.}

\item{method}{a character string indicating which correlation coefficient is to be computed.
Current version only support "pearson" correlation.}

\item{GPUidx}{is the integer index of GPU(device) starting from 0. Default is 0.}

\item{...}{not used.}
}
\value{
correlation matrix
}
\description{
Fast calculate correlation matrix on GPU
(the idea is from \pkg{WGCNA} fast function for pearson correlations)
}
\examples{
A <- matrix(rnorm(100, mean = 5, sd = 10), ncol = 10)
B <- matrix(rnorm(200, mean = 10, sd = 100), ncol = 20)
gpucor <- GPUgetcor(A, B)
cpucor <- cor(A, B)
gpucor[1:10, 1:5]
cpucor[1:10, 1:5]
}
\author{
Beibei Jiang \email{beibei_jiang@psych.mpg.de}
}
