% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\name{gpuMM}
\alias{gpuMM}
\title{Calculate matrix product on GPU: calling cublas library}
\usage{
gpuMM(A, B, m, k, n, idx = 0)
}
\arguments{
\item{A}{a matrix or a data.frame.}

\item{B}{a matrix or a data.frame.}

\item{m}{number of rows in A.}

\item{k}{number of columns in A. It should also be the number of rows in B.}

\item{n}{number of columns in B}

\item{idx}{the index of GPU(device). The default is 0.}
}
\value{
a product matrix which has m rows and n columns.
}
\description{
Calculate matrix product on GPU: calling cublas library
}
\examples{
set.seed(123)
m <- 100
k <- 200
n <- 100
A <- matrix(rnorm(m*k), ncol = k)
B <- matrix(rnorm(k*n), ncol = n)
gpuC <- gpuMM(A, B, m, k, n)
cpuC <- A\%*\%B
gpuC[1:6, 1:5]
cpuC[1:6, 1:5]
}
