% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{ZtoP}
\alias{ZtoP}
\title{Convert Z-score to correponding p-value}
\usage{
ZtoP(z.score, ...)
}
\arguments{
\item{z.score}{Z-score(s) (either scalar or vector).}

\item{...}{not used.}
}
\value{
corresponding p-value(s).
}
\description{
Convert Z score to correponding p-values
}
\note{
Due to the limitation of IEEE number representation in doubles,
any Z score over 37.51929999999999765 will be converted to a p-value of 1e-309.
}
\author{
Beibei Jiang \email{beibei_jiang@psych.mpg.de} and Benno Pütz \email{puetz@psych.mpg.de}
}
